Rename `GameplayLeaderboard` to `HUDLeaderboard` and move to proper package
